Update security, connection and player logic
Updated the security configuration by removing hardcoded password for Redis.
Fixed type mismatch in 'NewConnID' method in 'ConnManager'.
Updated 'Player' struct to contain 'Conn' and 'Data' and added 'Initialization' function to structure player data.
Changed the return type of 'NewConnID' in 'IConnManager' from int64 to int.
Added password encryption in 'Write Account' and improved data marshalling in 'SetPlayerInfo'.
Enhanced player info retrieval and setting in 'GetPlayerInfo' and 'SetPlayerInfo'.
Refactored connect and disconnect hooks in 'ConnManager', and created separate locks for connections and closed connection IDs.
Tweaked 'DoMsgHandler' to check if the account data is null instead of the player.
Updated the concurrent client connections logic for both socket and websockets.
These changes enhance the server's security, connection management, and data handling efficiency.

+++ iface/iconnmanager.go
@@ -3,7 +3,7 @@ package iface
 // 连接管理器
 type IConnManager interface {
 	// 生成一个新的连接ID
-	NewConnID() int64
+	NewConnID() int
 	// 添加连接
 	Add(conn IConnection)
 	// 删除连接
